@@ -3,19 +3,19 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@reduxjs/toolkit": "^1.6.2",
+		"@reduxjs/toolkit": "^1.8.0",
 		"@testing-library/jest-dom": "^4.2.4",
 		"@testing-library/react": "^9.5.0",
 		"@testing-library/user-event": "^7.2.1",
 		"chance": "^1.1.6",
-		"moment": "^2.27.0",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
-		"react-redux": "^7.2.0",
-		"react-router-dom": "^6.0.0",
-		"react-scripts": "4.0.0",
-		"redux": "^4.0.5",
-		"typescript": "^4.1.2"
+		"moment": "^2.29.0",
+		"react": "^18.0.0",
+		"react-dom": "^18.0.0",
+		"react-redux": "^8.0.0",
+		"react-router-dom": "^6.4.0",
+		"react-scripts": "^5.0.0",
+		"redux": "^4.2.0",
+		"typescript": "^4.8.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -39,11 +39,10 @@
 		]
 	},
 	"devDependencies": {
-		"eslint": "^7.32.0",
-		"eslint-config-standard": "^16.0.3",
-		"eslint-plugin-import": "^2.25.2",
+		"eslint": "^8.0.1",
+		"eslint-config-standard": "^17.0.0",
+		"eslint-plugin-import": "^2.26.0",
 		"eslint-plugin-node": "^11.1.0",
-		"eslint-plugin-promise": "^5.1.1",
-		"eslint-plugin-react": "^7.26.1"
+		"eslint-plugin-react": "^7.31.0"
 	}
 }
